requirements.txt
fastapi
motor
uvicorn
pydantic
bson
python-dotenv
